requirements.txt
gpiozero
pyzmq
Adafruit-IO
smtplib
someipy